+++ src/games/index.ts
@@ -37,6 +37,7 @@ import { StringsGame, IStringsState } from "./strings";
 import { WitchGame, IWitchState } from "./witch";
 import { ComplicaGame, IComplicaState } from "./complica";
 import { PigsGame, IPigsState } from "./pigs";
+import { GardenGame, IGardenState } from "./garden";
 
 export {
     APGamesInformation, GameBase, GameBaseSimultaneous, IAPGameState,
@@ -76,6 +77,7 @@ export {
     WitchGame, IWitchState,
     ComplicaGame, IComplicaState,
     PigsGame, IPigsState,
+    GardenGame, IGardenState,
 };
 
 const games = new Map<string, typeof AmazonsGame | typeof BlamGame | typeof CannonGame |
@@ -89,10 +91,11 @@ const games = new Map<string, typeof AmazonsGame | typeof BlamGame | typeof Cann
                               typeof ZolaGame | typeof MonkeyQueenGame | typeof DipoleGame |
                               typeof AlfredsWykeGame | typeof RealmGame | typeof ACityGame |
                               typeof FanoronaGame | typeof FocusGame | typeof StringsGame |
-                              typeof WitchGame | typeof ComplicaGame | typeof PigsGame
+                              typeof WitchGame | typeof ComplicaGame | typeof PigsGame |
+                              typeof GardenGame
                 >();
 // Manually add each game to the following array
-[AmazonsGame, BlamGame, CannonGame, MchessGame, HomeworldsGame, EntropyGame, VolcanoGame, MvolcanoGame, ChaseGame, AbandeGame, CephalopodGame, LinesOfActionGame, PikemenGame, OrdoGame, AttangleGame, AccastaGame, EpamGame, TaijiGame, BreakthroughGame, FabrikGame, ManalathGame, UrbinoGame, FendoGame, ArchimedesGame, ZolaGame, MonkeyQueenGame, DipoleGame, AlfredsWykeGame, RealmGame, ACityGame, FanoronaGame, FocusGame, StringsGame, WitchGame, ComplicaGame, PigsGame].forEach((g) => {
+[AmazonsGame, BlamGame, CannonGame, MchessGame, HomeworldsGame, EntropyGame, VolcanoGame, MvolcanoGame, ChaseGame, AbandeGame, CephalopodGame, LinesOfActionGame, PikemenGame, OrdoGame, AttangleGame, AccastaGame, EpamGame, TaijiGame, BreakthroughGame, FabrikGame, ManalathGame, UrbinoGame, FendoGame, ArchimedesGame, ZolaGame, MonkeyQueenGame, DipoleGame, AlfredsWykeGame, RealmGame, ACityGame, FanoronaGame, FocusGame, StringsGame, WitchGame, ComplicaGame, PigsGame, GardenGame].forEach((g) => {
     if (games.has(g.gameinfo.uid)) {
         throw new Error("Another game with the UID '" + g.gameinfo.uid + "' has already been used. Duplicates are not allowed.");
     }
@@ -175,6 +178,8 @@ export const GameFactory = (game: string, ...args: any[]): GameBase|GameBaseSimu
             return new ComplicaGame(...args);
         case "pigs":
             return new PigsGame(...args);
+        case "garden":
+            return new GardenGame(...args);
     }
     return;
 }