+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * New game: Witch Stones
 * New game: Complica
 * New game: Robo Battle Pigs
+<<<<<<< HEAD
+=======
+* New game: Wizard's Garden
+>>>>>>> develop
 * Added `experimental` flag so games can be excluded in production, even if updated game code gets deployed to production.
 
 ### Fixed